move BuildspecParser to buildtest/buildsystem/parser.py Move test/buuildsystem/schema --> tests/schema to map directory structure relative in "buildtest" Remove .yaml extension from load_schema during re.search
