+++ tests/buildsystem/test_base.py
@@ -5,7 +5,7 @@ BuildspecParser: testing functions
 import pytest
 import os
 
-from buildtest.buildsystem.base import BuildspecParser
+from buildtest.buildsystem.parser import BuildspecParser
 
 testroot = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
